+++ src/core/tests/MultiHashMap.test.ts
@@ -1,4 +1,4 @@
-import { assertEquals } from "../../testUtils/mod.ts";
+import { assertEquals } from "../../testing/mod.ts";
 import { MultiHashMap } from "../mod.ts";
 
 Deno.test("MultiHashMap", () => {
