+++ src/serial/tests/index.test.ts
@@ -24,7 +24,7 @@ import {
   Serializer,
 } from "../mod.ts";
 
-import { assertEquals, snapshot } from "../../testUtils/mod.ts";
+import { assertEquals, snapshot } from "../../testing/mod.ts";
 
 // deno-fmt-ignore
 const lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
